Move getConfirmCancelButtonRow to utils.ts
